+++ pysatMadrigal/instruments/methods/general.py
@@ -1,3 +1,8 @@
+#!/usr/bin/env python
+# Full license can be found in License.md
+# Full author list can be found in .zenodo.json file
+# DOI:10.5281/zenodo.3824979
+# ----------------------------------------------------------------------------
 # -*- coding: utf-8 -*-.
 """General routines for integrating CEDAR Madrigal instruments into pysat.
 
@@ -32,6 +37,459 @@ def cedar_rules():
     ackn = "".join(["Contact the PI when using this data, in accordance ",
                     "with the CEDAR 'Rules of the Road'"])
     return ackn
+
+
+def known_madrigal_inst_codes(pandas_format=None):
+    """Supplies known Madrigal instrument codes with a brief description.
+
+    Parameters
+    ----------
+    pandas_format : bool or NoneType
+       Separate instrument codes by time-series (True) or multi-dimensional
+       data types (False) if a boolean is supplied, or supply all if NoneType
+       (default=None)
+
+    Returns
+    -------
+    inst_codes : dict
+        Dictionary with string instrument code values as keys and a brief
+        description of the corresponding instrument as the value.
+
+    """
+
+    time_series = {'120': 'Interplanetary Mag Field and Solar Wind',
+                   '210': 'Geophysical Indicies', '211': 'AE Index',
+                   '212': 'DST Index', '170': 'POES Spacecraft Particle Flux',
+                   '180': 'DMSP-Auroral Boundary Index',
+                   '8100': 'Defense Meteorological Satellite Program',
+                   '8105': 'Van Allen Probes', '8400': 'Jason/Topex Ocean TEC',
+                   '8250': 'Jicamarca Magnetometer',
+                   '8255': 'Piura Magnetometer',
+                   '8300': 'Sodankyla Magnetometer',
+                   '7800': 'Green Bank Telescope'}
+    multi_dim = {'10': 'Jicamarca ISR', '20': 'Arecibo ISR Linefeed',
+                 '21': 'Arecibo ISR Gregorian',
+                 '22': 'Arecibo ISR Velocity Vector',
+                 '25': 'MU ISR', '30': 'Millstone Hill ISR',
+                 '31': 'Millstone Hill UHF Steerable Antenna',
+                 '32': 'Millstone Hill UHF Zenith Antenna',
+                 '40': 'St. Santin ISR', '41': 'St. Santin Nançay Receiver',
+                 '42': 'St. Santin Mende Receiver',
+                 '43': 'St. Santin Monpazier Receiver',
+                 '45': 'Kharkov Ukraine ISR', '50': 'Chatanika ISR',
+                 '53': 'ISTP Irkutsk Radar', '57': 'UK Malvern ISR',
+                 '61': 'Poker Flat ISR', '70': 'EISCAT combined ISRs',
+                 '71': 'EISCAT Kiruna UHF ISR', '72': 'EISCAT Tromsø UHF ISR',
+                 '73': 'EISCAT Sodankylä UHF ISR',
+                 '74': 'EISCAT Tromsø VHF ISR', '75': 'EISCAT Kiruna VHF ISR',
+                 '76': 'EISCAT Sodankylä VHF ISR', '80': 'Sondrestrom ISR',
+                 '85': 'ALTAIR ISR', '91': 'Resolute Bay North ISR',
+                 '92': 'Resolute Bay Canada ISR',
+                 '95': 'EISCAT Svalbard ISR Longyearbyen',
+                 '100': 'QuJing ISR', '310': 'TGCM/TIGCM model',
+                 '311': 'AMIE Model', '312': 'USU-TDIM Model',
+                 '320': 'Solar sd Tides', '321': 'Lunar sd Tides',
+                 '322': 'GSWM model', '820': 'Halley HF Radar',
+                 '830': 'Syowa Station HF Radar', '845': 'Kapuskasing HF Radar',
+                 '861': 'Saskatoon HF Radar', '870': 'Goose Bay HF Radar',
+                 '900': 'Hankasalmi HF Radar', '910': 'Stokkseyri HF Radar',
+                 '911': 'Pykkvibaer HF Radar', '1040': 'Arecibo MST Radar',
+                 '1140': 'Poker Flat MST Radar',
+                 '1180': 'SOUSY Svalbard MST Radar Longyearbyen',
+                 '1210': 'Scott Base MF Radar',
+                 '1215': 'Davis Antarctica MF radar',
+                 '1220': 'Mawson MF Radar', '1221': 'Rothera MF radar',
+                 '1230': 'Christchurch MF Radar',
+                 '1240': 'Adelaide MF Radar', '1245': 'Rarotonga MF radar',
+                 '1254': 'Tirunelveli MF radar', '1270': 'Kauai MF radar',
+                 '1275': 'Yamagawa MF radar', '1285': 'Platteville MF radar',
+                 '1310': 'Wakkanai MF radar', '1320': 'Collm LF Radar',
+                 '1340': 'Saskatoon MF Radar',
+                 '1375': 'The Poker Flat MF radar', '1390': 'Tromsø MF Radar',
+                 '1395': 'Syowa MF Radar', '1400': 'Halley MF Radar',
+                 '13': 'JASMET Jicamarca All-Sky Specular Meteor Radar',
+                 '1539': 'Ascension Island Meteor Radar',
+                 '1540': 'Rothera Meteor Radar',
+                 '1560': 'Atlanta meteor Radar', '1620': 'Durham meteor Radar',
+                 '1750': 'Obninsk meteor radar', '1775': 'Esrange meteor radar',
+                 '1780': 'Wuhan meteor radar', '1781': 'Mohe meteor radar',
+                 '1782': 'Beijing meteor radar', '1783': 'Sanya meteor radar',
+                 '1784': 'South Pole meteor radar',
+                 '1785': 'Southern Argentina Agile Meteor Radar',
+                 '1786': 'Cachoeira Paulista Meteor Radar',
+                 '1787': 'Buckland Park Meteor Radar',
+                 '1788': 'Kingston Meteor Radar', '1790': 'Andes Meteor Radar',
+                 '1791': 'Southern Cross Meteor Radar',
+                 '1792': 'Las Campanas Meteor Radar',
+                 '1793': 'CONDOR multi-static meteor radar system',
+                 '2090': 'Christmas Island ST/MEDAC Radar',
+                 '2200': 'Platteville ST/MEDAC Radar',
+                 '2550': 'ULowell Digisonde MLH Radar',
+                 '2890': 'Sondre Stromfjord Digisonde',
+                 '2900': 'Sodankylä Ionosonde (SO166)',
+                 '2930': 'Qaanaaq Digisonde ST/MEDAC Radars',
+                 '2950': 'EISCAT Tromsø Dynasonde',
+                 '2951': 'EISCAT Svalbard Dynasonde',
+                 '2952': 'IRF Dynasonde at EISCAT site Kiruna',
+                 '5000': 'South Pole Fabry-Perot', '5005': 'Palmer Fabry Perot',
+                 '5015': 'Arrival Heights Fabry-Perot',
+                 '5020': 'Halley Fabry-Perot',
+                 '5060': 'Mount John Fabry-Perot',
+                 '5140': 'Fabry-Perot Arequipa',
+                 '5145': 'Fabry-Perot Jicamarca', '5150': 'Fabry-Perot Mobile',
+                 '5160': 'Arecibo Fabry-Perot',
+                 '5190': 'Kitt Peak H-alpha Fabry-Perot',
+                 '5240': 'Fritz Peak Fabry-Perot',
+                 '5292': 'Ann Arbor Fabry-Perot',
+                 '5300': 'Peach Mountain Fabry-Perot',
+                 '5340': 'Millstone Hill Fabry-Perot',
+                 '5360': 'Millstone Hill High-Res Fabry-Perot',
+                 '5370': 'Arecibo Imaging Doppler Fabry-Perot',
+                 '5380': 'Culebra Fabry-Perot',
+                 '5430': 'Watson Lake Fabry-Perot',
+                 '5460': 'College Fabry-Perot',
+                 '5465': 'Poker Flat all-sky scanning Fabry-Perot',
+                 '5470': 'Fort Yukon Fabry-Perot',
+                 '5475': 'Poker Flat Fabry-Perot',
+                 '5480': 'Sondre Stromfjord Fabry-Perots',
+                 '5510': 'Inuvik NWT Fabry-Perot',
+                 '5535': 'Resolute Bay Fabry-Perot',
+                 '5540': 'Thule Fabry-Perot', '5545': 'Cariri Brazil FPI',
+                 '5546': 'Cajazeiras Brazil FPI',
+                 '5547': 'Pisgah Astronomical Research FPI',
+                 '5548': 'Urbana Atmospheric Observatory FPI',
+                 '5549': 'Kirtland Airforce Base FPI',
+                 '5550': 'Virginia Tech FPI',
+                 '5551': 'Peach Mountain (MiniME) FPI',
+                 '5552': 'Merihill Peru FPI', '5553': 'Nazca Peru FPI',
+                 '5554': 'Eastern Kentucky FPI',
+                 '5600': 'Jang Bogo Station FPI',
+                 '5700': 'South Pole Michelson Interferometer',
+                 '5720': 'Daytona Beach Michelson Interferometer',
+                 '5860': 'Stockholm IR Michelson',
+                 '5900': 'Sondrestrom Michelson Interferometer',
+                 '5950': 'Resolute Bay Michelson Interferometer',
+                 '5980': 'Eureka Michelson Interferometer',
+                 '6205': 'Arecibo Potassium [K] lidar',
+                 '6206': 'Arecibo Sodium [Na] lidar',
+                 '6300': 'CEDAR lidar', '6320': 'Colorado State sodium lidar',
+                 '6330': 'Rayleigh lidar at the ALO - USU/CASS',
+                 '6340': 'Andes Na T/W Lidar', '6350': 'ALOMAR Sodium Lidar',
+                 '6360': 'CU STAR Sodium Lidar', '6370': 'USU Na lidar',
+                 '6380': 'Poker Flat lidar', '7190': 'USU CCD Imager',
+                 '7192': 'USU Advanced Mesospheric Temperature Mapper',
+                 '7200': 'BU Millstone All-Sky Imager',
+                 '7201': 'BU Arecibo All-Sky Imager',
+                 '7202': 'BU Asiago All-Sky Imager',
+                 '7203': 'BU El Leoncito All-Sky Imager',
+                 '7204': 'BU McDonald All-Sky Imager',
+                 '7205': 'BU Rio Grande All-Sky Imager',
+                 '7206': 'BU Jicamarca All-Sky Imager', '7240': 'MIO',
+                 '7580': 'All-sky cameras at Qaanaaq',
+                 '11': 'Jicamarca Bistatic Radar', '840': 'JULIA',
+                 '3000': 'ARL UT TBB Receiver',
+                 '7600': 'Chelmsford HS Ozone Radiometer',
+                 '7602': 'Lancaster UK Ozone Radiometer',
+                 '7603': 'Bridgewater MA Ozone Radiometer',
+                 '7604': 'Union College Ozone Radiometer',
+                 '7605': 'UNC Greensboro Ozone Radiometer',
+                 '7606': 'Lynnfield HS Ozone Radiometer',
+                 '7607': 'Alaska Pacific Ozone Radiometer',
+                 '7608': 'Hermanus SA Ozone Radiometer',
+                 '7609': 'Sanae Antarctic Ozone Radiometer',
+                 '7610': 'Sodankylä Ozone Radiometer',
+                 '7611': 'Lancaster2 UK Ozone Radiometer',
+                 '7612': 'Haystack Ridge Ozone Radiometer',
+                 '7613': 'Haystack NUC3 8-channel Ozone Radiometer',
+                 '7614': 'Fairbanks Ozone Radiometer',
+                 '8001': 'South Pole Scintillation Receiver',
+                 '8000': 'World-wide GNSS Receiver Network',
+                 '8002': 'McMurdo Scintillation Receiver',
+                 '8010': 'GNSS Scintillation Network',
+                 '3010': 'Davis Czerny-Turner Scanning Spectrophotometer',
+                 '3320': 'Wuppertal (DE) Czerny-Turner OH Grating Spectrometer',
+                 '4470': 'Poker Flat 4 Channel Filter Photometer',
+                 '4473': 'Fort Yukon 4 Channel Filter Photometer',
+                 '4480': 'Arecibo red line photometer',
+                 '4481': 'Arecibo green line photometer',
+                 '7191': 'USU Mesospheric Temperature Mapper'}
+
+    if pandas_format is None:
+        inst_codes = dict(**time_series, **multi_dim)
+    elif pandas_format:
+        inst_codes = time_series
+    else:
+        inst_codes = multi_dim
+
+    return inst_codes
+
+
+def madrigal_file_format_str(inst_code, strict=False, verbose=True):
+    """Supplies known Madrigal instrument codes with a brief description.
+
+    Parameters
+    ----------
+    inst_code : int
+        Madrigal instrument code as an integer
+    strict : bool
+        If True, returns only file formats that will definitely not have a
+        problem being parsed by pysat.  If False, will return any file format.
+        (default=False)
+    verbose : bool
+        If True raises logging warnings, if False does not log any warnings.
+        (default=True)
+
+    Returns
+    -------
+    fstr : str
+        File formatting string that may or may not be parsable by pysat
+
+    Raises
+    ------
+    ValueError
+        If file formats with problems would be returned and `strict` is True.
+
+    Note
+    ----
+    File strings that have multiple '*' wildcards typically have several
+    experiment types and require a full pysat Instrument to properly manage
+    these types.
+
+    """
+
+    if not isinstance(inst_code, int):
+        inst_code = int(inst_code)
+
+    format_str = {
+        120: 'imf{{year:02d}}{{month:02d}}{{day:02d}}g.{{version:03d}}.',
+        210: 'geo{{year:02d}}{{month:02d}}{{day:02d}}g.{{version:03d}}.',
+        211: 'aei{{year:02d}}{{month:02d}}{{day:02d}}g.{{version:03d}}.',
+        212: 'dst{{year:02d}}{{month:02d}}{{day:02d}}g.{{version:03d}}.',
+        170: 'pfx{{year:02d}}{{month:02d}}{{day:02d}}g.{kindat}.',
+        180: 'dmp{{year:04d}}{{month:02d}}{{day:02d}}.{{version:03d}}.',
+        8100: 'dms*_{{year:04d}}{{month:02d}}{{day:02d}}_*.{{version:03d}}.',
+        8105: 'van_allen_{{year:04d}}_{{month:02d}}.{{version:03d}}.',
+        8400: '???{{year:04d}}{{month:02d}}{{day:02d}}j*.{{version:03d}}.',
+        8250: 'jic{{year:04d}}{{month:02d}}{{day:02d}}_mag.{{version:03d}}.',
+        8255: 'pmt*.',
+        8300: 'smt*.',
+        7800: 'gbt{{year:04d}}{{month:02d}}{{day:02d}}.{{version:03d}}.',
+        10: 'jro{{year:04d}}{{month:02d}}{{day:02d}}*.{{version:03d}}.',
+        20: 'aro*{{year:02d}}{{month:02d}}{{day:02d}}a.{{version:03d}}.',
+        21: 'aro*{{year:02d}}{{month:02d}}{{day:02d}}*g.{{version:03d}}.',
+        22: 'ar?*{{year:02d}}{{month:02d}}{{day:02d}}*.{{version:03d}}.',
+        25: 'mui{{year:02d}}{{month:02d}}{{day:02d}}?.{{version:03d}}.',
+        30: 'mlh{{year:02d}}{{month:02d}}{{day:02d}}?.{{version:03d}}.',
+        31: 'mlh{{year:02d}}{{month:02d}}{{day:02d}}?.{{version:03d}}.',
+        32: 'mlh{{year:02d}}{{month:02d}}{{day:02d}}.{{version:03d}}.',
+        40: 'sts{{year:02d}}{{month:02d}}{{day:02d}}g.{{version:03d}}.',
+        41: 'sts{{year:02d}}{{month:02d}}{{day:02d}}g.{{version:03d}}.',
+        42: 'sts{{year:02d}}{{month:02d}}{{day:02d}}g.{{version:03d}}.',
+        43: 'sts{{year:02d}}{{month:02d}}{{day:02d}}g.{{version:03d}}.',
+        45: 'kha{{year:02d}}{{month:02d}}{{day:02d}}g.{{version:03d}}.',
+        50: 'cht{{year:02d}}{{month:02d}}{{day:02d}}g.{kindat}.',
+        53: 'ist{{year:02d}}{{month:02d}}{{day:02d}}g.{{version:03d}}.',
+        57: 'mlv{{year:02d}}{{month:02d}}{{day:02d}}g.{{version:03d}}.',
+        61: 'pfa{{year:02d}}{{month:02d}}{{day:02d}}g.{{version:03d}}.',
+        70: 'MAD????_{year:04d}}-{{month:02d}}-{{day:02d}}_*.',
+        71: 'MAD????_{year:04d}}-{{month:02d}}-{{day:02d}}_*@kir.',
+        72: 'MAD????_{year:04d}}-{{month:02d}}-{{day:02d}}_*@uhf.',
+        73: 'MAD????_{year:04d}}-{{month:02d}}-{{day:02d}}_*@sod.',
+        74: 'MAD????_{year:04d}}-{{month:02d}}-{{day:02d}}_*@vhf.',
+        75: 'MAD????_{year:04d}}-{{month:02d}}-{{day:02d}}_*@vkrv*.',
+        76: 'MAD????_{year:04d}}-{{month:02d}}-{{day:02d}}_*@sdv*.',
+        80: 'son{{year:02d}}{{month:02d}}{{day:02d}}.{{version:03d}}.',
+        85: 'ALT{{year:02d}}{{month:02d}}{{day:02d}}_*.',
+        91: 'ran{{year:02d}}{{month:02d}}{{day:02d}}.{{version:03d}}.',
+        92: 'ras{{year:02d}}{{month:02d}}{{day:02d}}.{{version:03d}}.',
+        95: 'MAD????_{year:04d}}-{{month:02d}}-{{day:02d}}_*@esr.',
+        100: 'MAD????_{year:04d}}-{{month:02d}}-{{day:02d}}_*@quj.',
+        310: 'gcm*.',
+        311: 'ami*.',
+        312: 'tdi*.',
+        320: 'sdt*.',
+        321: 'sdl*.',
+        322: 'gsw*.',
+        820: 'hhf*.',
+        830: 'syf*.',
+        845: 'khf*.',
+        861: 'shf*.',
+        870: 'gbf*.',
+        900: 'fhf*.',
+        910: 'whf*.',
+        911: 'ehf*.',
+        1040: 'arm{{year:02d}}{{month:02d}}{{day:02d}}g.{kindat}.',
+        1140: 'pkr{{year:02d}}{{month:02d}}{{day:02d}}g.{kindat}.',
+        1180: 'ssr*.',
+        1210: 'sbf{{year:02d}}{{month:02d}}{{day:02d}}g.{kindat}.',
+        1215: 'dav{{year:02d}}{{month:02d}}{{day:02d}}g.{kindat}.',
+        1220: 'maf{{year:02d}}{{month:02d}}{{day:02d}}g.{kindat}.',
+        1221: 'rth{{year:02d}}{{month:02d}}{{day:02d}}g.{kindat}.',
+        1230: 'ccf{{year:02d}}{{month:02d}}{{day:02d}}g.{kindat}.',
+        1240: 'adf{{year:02d}}{{month:02d}}{{day:02d}}g.{kindat}.',
+        1245: 'rtg{{year:02d}}{{month:02d}}{{day:02d}}g.{kindat}.',
+        1254: 'tyr{{year:02d}}{{month:02d}}{{day:02d}}g.{kindat}.',
+        1270: 'kau{{year:02d}}{{month:02d}}{{day:02d}}g.{kindat}.',
+        1275: 'yam{{year:02d}}{{month:02d}}{{day:02d}}g.{kindat}.',
+        1285: 'plr{{year:02d}}{{month:02d}}{{day:02d}}g.{kindat}.',
+        1310: 'wak{{year:02d}}{{month:02d}}{{day:02d}}g.{kindat}.',
+        1320: 'cof{{year:02d}}{{month:02d}}{{day:02d}}g.{kindat}.',
+        1340: 'saf{{year:02d}}{{month:02d}}{{day:02d}}g.{kindat}.',
+        1375: 'rpk{{year:02d}}{{month:02d}}{{day:02d}}g.{kindat}.',
+        1390: 'trf{{year:02d}}{{month:02d}}{{day:02d}}g.{kindat}.',
+        1395: 'sym_{{year:04d}}{{month:02d}}{{day:02d}}.{{version:03d}}.',
+        1400: 'hmf_{{year:04d}}{{month:02d}}{{day:02d}}.{{version:03d}}.',
+        13: 'D{{year:04d}}{{month:02d}}*.',
+        1539: 'asc{{year:02d}}{{month:02d}}{{day:02d}}g.{kindat}.',
+        1540: 'rmr_{{year:04d}}{{month:02d}}{{day:02d}}.{{version:03d}}.',
+        1560: 'atm{{year:02d}}{{month:02d}}{{day:02d}}g.{kindat}.',
+        1620: 'dum{{year:02d}}{{month:02d}}{{day:02d}}g.{{version:03d}}.',
+        1750: 'obn{{year:02d}}{{month:02d}}{{day:02d}}g.{kindat}.',
+        1775: 'emr{{year:02d}}{{month:02d}}{{day:02d}}g.{kindat}.',
+        1780: 'wmr*.',
+        1781: 'mmr*.',
+        1782: 'bmr*.',
+        1783: 'smr*.',
+        1784: 'som{{year:02d}}{{month:02d}}{{day:02d}}.{{version:03d}}.',
+        1785: 'amr{{year:04d}}{{month:02d}}{{day:02d}}.{{version:03d}}.',
+        1786: 'cpr_{{year:04d}}{{month:02d}}{{day:02d}}.{{version:03d}}.',
+        1787: 'bpr_{{year:04d}}{{month:02d}}{{day:02d}}.{{version:03d}}.',
+        1788: 'kgr_{{year:04d}}{{month:02d}}{{day:02d}}.{{version:03d}}.',
+        1790: 'ame*.',
+        1791: 'sco*.',
+        1792: 'lcm*.',
+        1793: 'alo{{year:04d}}{{month:02d}}{{day:02d}}_{{version:03d}}.',
+        2090: 'cia{{year:02d}}{{month:02d}}{{day:02d}}g.{kindat}.',
+        2200: 'pla{{year:02d}}{{month:02d}}{{day:02d}}g.{kindat}.',
+        2550: 'uld*.',
+        2890: 'ssd{{year:02d}}{{month:02d}}{{day:02d}}g.{kindat}.',
+        2900: 'sdi*.',
+        2930: 'qad{{year:02d}}{{month:02d}}{{day:02d}}g.{kindat}.',
+        2950: 'trd*.',
+        2951: 'lrd*.',
+        2952: 'krd*.',
+        5000: 'spf{{year:02d}}{{month:02d}}{{day:02d}}g.{kindat}.',
+        5005: 'pfi{{year:04d}}{{month:02d}}{{day:02d}}.',
+        5015: 'ahf{{year:02d}}{{month:02d}}{{day:02d}}g.{kindat}.',
+        5020: 'hfp{{year:02d}}{{month:02d}}{{day:02d}}g.{kindat}.',
+        5060: 'mjf{{year:02d}}{{month:02d}}{{day:02d}}g.{kindat}.',
+        5140: 'aqf{{year:02d}}{{month:02d}}{{day:02d}}g.{kindat}.',
+        5145: 'jfp{{year:04d}}{{month:02d}}{{day:02d}}_*.{{version:03d}}.',
+        5150: 'mfp{{year:04d}}{{month:02d}}{{day:02d}}_*.{{version:03d}}.',
+        5160: 'afp{{year:02d}}{{month:02d}}{{day:02d}}g.{kindat}.',
+        5190: 'kha{{year:02d}}{{month:02d}}{{day:02d}}g.{kindat}.',
+        5240: 'fpf{{year:02d}}{{month:02d}}{{day:02d}}g.{kindat}.',
+        5292: 'aaf{{year:02d}}{{month:02d}}{{day:02d}}g.{kindat}.',
+        5300: 'pfp{{year:02d}}{{month:02d}}{{day:02d}}g.{kindat}.',
+        5340: 'mfp{{year:02d}}{{month:02d}}{{day:02d}}g.{{version:03d}}.?.',
+        5360: 'kfp{{year:02d}}{{month:02d}}{{day:02d}}g*.',
+        5370: 'aif{{year:02d}}{{month:02d}}{{day:02d}}g*.',
+        5380: 'clf{{year:02d}}{{month:02d}}{{day:02d}}g.{kindat}.',
+        5430: 'wfp{{year:02d}}{{month:02d}}{{day:02d}}g.{kindat}.',
+        5460: 'cfp{{year:02d}}{{month:02d}}{{day:02d}}g.{kindat}.',
+        5465: 'pkf{{year:02d}}{{month:02d}}{{day:02d}}*.',
+        5470: 'FYU{{year:04d}}{{month:02d}}{{day:02d}}.',
+        5475: 'PKZ{{year:04d}}{{month:02d}}{{day:02d}}.',
+        5480: 'sfp{{year:02d}}{{month:02d}}{{day:02d}}g.{kindat}.',
+        5510: 'ikf{{year:02d}}{{month:02d}}{{day:02d}}g.{kindat}.',
+        5535: 'rfp{{year:02d}}{{month:02d}}{{day:02d}}g.{kindat}.',
+        5540: 'tfp{{year:02d}}{{month:02d}}{{day:02d}}g.{kindat}.',
+        5545: ''.join(['minime01_car_{{year:04d}}{{month:02d}}{{day:02d}}.',
+                       'cedar.{{version:03d}}.']),
+        5546: ''.join(['minime02_caj_{{year:04d}}{{month:02d}}{{day:02d}}.',
+                       'cedar.{{version:03d}}.']),
+        5547: ''.join(['minime06_par_{{year:04d}}{{month:02d}}{{day:02d}}.',
+                       'cedar.{{version:03d}}.']),
+        5548: ''.join(['minime02_uao_{{year:04d}}{{month:02d}}{{day:02d}}.',
+                       'cedar.{{version:03d}}.']),
+        5549: 'Kirtland Airforce Base FPI',
+        5550: ''.join(['minime09_vti_{{year:04d}}{{month:02d}}{{day:02d}}.',
+                       'cedar.{{version:03d}}.']),
+        5551: ''.join(['minime08_ann_{{year:04d}}{{month:02d}}{{day:02d}}.',
+                       'cedar.{{version:03d}}.']),
+        5552: 'Merihill Peru FPI',
+        5553: 'Nazca Peru FPI',
+        5554: ''.join(['minime07_euk_{{year:04d}}{{month:02d}}{{day:02d}}.',
+                       'cedar.{{version:03d}}.']),
+        5600: 'jbs_{{year:04d}}{{month:02d}}{{day:02d}}.{{version:03d}}.',
+        5700: 'spm{{year:02d}}{{month:02d}}{{day:02d}}g.{kindat}.',
+        5720: 'dbm{{year:02d}}{{month:02d}}{{day:02d}}g.{kindat}.',
+        5860: 'stm{{year:02d}}{{month:02d}}{{day:02d}}g.{kindat}.',
+        5900: 'sfm{{year:02d}}{{month:02d}}{{day:02d}}g.{kindat}.',
+        5950: 'rbm{{year:02d}}{{month:02d}}{{day:02d}}g.{kindat}.',
+        5980: 'eum{{year:02d}}{{month:02d}}{{day:02d}}g.{kindat}.',
+        6205: 'akl{{year:02d}}{{month:02d}}{{day:02d}}g.*.',
+        6206: 'asl{{year:04d}}{{month:02d}}{{day:02d}}.{{version:03d}}.',
+        6300: 'uil{{year:02d}}{{month:02d}}{{day:02d}}g.{kindat}.',
+        6320: 'Colorado State sodium lidar',
+        6330: 'usl{{year:02d}}{{month:02d}}{{day:02d}}g.{kindat}.',
+        6340: 'alo*.',
+        6350: 'nlo*.',
+        6360: 'cul*.',
+        6370: 'unl*.',
+        6380: 'pfl{{year:04d}}{{month:02d}}{{day:02d}}_{{cycle:03d}}.',
+        7190: 'usi*.',
+        7192: 'amp{{year:02d}}{{month:02d}}{{day:02d}}?.{{version:03d}}.',
+        7200: 'mhi{{year:04d}}{{month:02d}}{{day:02d}}.{kindat}.',
+        7201: 'aai{{year:04d}}{{month:02d}}{{day:02d}}.{kindat}.',
+        7202: 'abi{{year:04d}}{{month:02d}}{{day:02d}}.{kindat}.',
+        7203: 'eai{{year:04d}}{{month:02d}}{{day:02d}}.{kindat}.',
+        7204: 'mai{{year:04d}}{{month:02d}}{{day:02d}}.{kindat}.',
+        7205: 'rai{{year:04d}}{{month:02d}}{{day:02d}}.{kindat}.',
+        7206: 'jci{{year:04d}}{{month:02d}}{{day:02d}}.{kindat}.',
+        7240: 'mhi*.',
+        7580: 'qac*.',
+        11: 'j??*{{year:02d}}{{month:02d}}{{day:02d}}g.{{version:03d}}.',
+        840: 'jul{{year:04d}}{{month:02d}}{{day:02d}}_esf.{{version:03d}}.',
+        3000: 'utx*.',
+        8001: '????_?_??.gps_all.out.',
+        8000: '*{{year:02d}}{{month:02d}}{{day:02d}}*.{{version:03d}}.',
+        8002: '????_?_??.gps_all.out.',
+        8010: 'scin_{{year:04d}}{{month:02d}}{{day:02d}}.{{version:03d}}.',
+        3010: 'dvs{{year:02d}}{{month:02d}}{{day:02d}}g.{kindat}.',
+        3320: 'wup{{year:02d}}{{month:02d}}{{day:02d}}g.{kindat}.',
+        4470: 'p4p{{year:02d}}{{month:02d}}{{day:02d}}g.{kindat}.',
+        4473: 'y4p{{year:02d}}{{month:02d}}{{day:02d}}g.{kindat}.',
+        4480: 'arp{{year:04d}}{{month:02d}}{{day:02d}}.{{version:03d}}.',
+        4481: 'agp{{year:04d}}{{month:02d}}{{day:02d}}.{{version:03d}}.',
+        7191: 'mtm{{year:02d}}{{month:02d}}{{day:02d}}g.{kindat}.'}
+
+    # Warn if file format not available
+    msg = ""
+    if inst_code not in format_str.keys():
+        msg = "".join(["file format string not available for ",
+                       "instrument code {:d}: ".format(inst_code)])
+        fstr = "*."
+    else:
+        fstr = format_str[inst_code]
+
+    # Warn if file format has multiple '*' wildcards
+    num_wc = len(fstr.split("*"))
+    if num_wc >= 3:
+        msg = "".join(["file format string has multiple '*' ",
+                       "wildcards, may not be parsable by pysat"])
+    elif fstr.find('{{year') < 0 and fstr != "*.":
+        msg = "".join(["file format string missing date info, ",
+                       "may not be parsable by pysat"])
+    elif num_wc > 1:
+        nspec_sec = 0
+        for fsplit in fstr.split("*"):
+            if fsplit.find("}}") > 0 and fsplit.find("{{") >= 0:
+                nspec_sec += 1
+
+        if nspec_sec > 1:
+            msg = "".join(["file format string has '*' between formatting",
+                           " constraints, may not be parsable by pysat"])
+
+    if len(msg) > 0:
+        if strict:
+            raise ValueError(msg)
+        elif verbose:
+            logger.warning(msg)
+
+    fstr += "{file_type}"
+
+    return fstr
 
 
 def load(fnames, tag='', inst_id='', xarray_coords=None):
@@ -459,7 +917,7 @@ def download(date_array, inst_code=None, kindat=None, data_path=None,
     return
 
 
-def get_remote_filenames(inst_code=None, kindat=None, user=None, password=None,
+def get_remote_filenames(inst_code=None, kindat='', user=None, password=None,
                          web_data=None, url="http://cedar.openmadrigal.org",
                          start=dt.datetime(1900, 1, 1), stop=dt.datetime.now(),
                          date_array=None):
@@ -470,10 +928,10 @@ def get_remote_filenames(inst_code=None, kindat=None, user=None, password=None,
     inst_code : str or NoneType
         Madrigal instrument code(s), cast as a string.  If multiple are used,
         separate them with commas. (default=None)
-    kindat : str or NoneType
+    kindat : str
         Madrigal experiment code(s), cast as a string.  If multiple are used,
         separate them with commas.  If not supplied, all will be returned.
-        (default=None)
+        (default='')
     data_path : str or NoneType
         Path to directory to download data to. (default=None)
     user : str or NoneType
@@ -487,10 +945,12 @@ def get_remote_filenames(inst_code=None, kindat=None, user=None, password=None,
         (default=None)
     url : str
         URL for Madrigal site (default='http://cedar.openmadrigal.org')
-    start : dt.datetime
-        Starting time for file list (defaults to 01-01-1900)
-    stop : dt.datetime
-        Ending time for the file list (defaults to time of run)
+    start : dt.datetime or NoneType
+        Starting time for file list, None reverts to default
+        (default=dt.datetime(1900, 1, 1))
+    stop : dt.datetime or NoneType
+        Ending time for the file list, None reverts to default
+        (default=dt.datetime.utcnow())
     date_array : dt.datetime or NoneType
         Array of datetimes to download data for. The sequence of dates need not
         be contiguous and will be used instead of start and stop if supplied.
@@ -522,7 +982,7 @@ def get_remote_filenames(inst_code=None, kindat=None, user=None, password=None,
 
     _check_madrigal_params(inst_code=inst_code, user=user, password=password)
 
-    if kindat is None:
+    if kindat in ['', '*']:
         kindat = []
     else:
         kindat = [int(kk) for kk in kindat.split(",")]
@@ -534,9 +994,18 @@ def get_remote_filenames(inst_code=None, kindat=None, user=None, password=None,
                 date_array))
         start = date_array.min()
         stop = date_array.max()
+
+    # If NoneType was supplied for start or stop, set to defaults
+    if start is None:
+        start = dt.datetime(1900, 1, 1)
+
+    if stop is None:
+        stop = dt.datetime.utcnow()
+
     # If start and stop are identical, increment
     if start == stop:
         stop += dt.timedelta(days=1)
+
     # Open connection to Madrigal
     if web_data is None:
         web_data = madrigalWeb.MadrigalData(url)
@@ -550,11 +1019,12 @@ def get_remote_filenames(inst_code=None, kindat=None, user=None, password=None,
 
     # Iterate over experiments to grab files for each one
     files = list()
-    logger.info("Found {:d} Madrigal experiments".format(len(exp_list)))
+    istr = "Found {:d} Madrigal experiments between {:s} and {:s}".format(
+        len(exp_list), start.strftime('%d %B %Y'), stop.strftime('%d %B %Y'))
+    logger.info(istr)
     for exp in exp_list:
         if good_exp(exp, date_array=date_array):
             file_list = web_data.getExperimentFiles(exp.id)
-
             if len(kindat) == 0:
                 files.extend(file_list)
             else:
@@ -718,8 +1188,12 @@ def list_remote_files(tag, inst_id, inst_code=None, kindats=None, user=None,
 
     # Parse these filenames to grab out the ones we want
     logger.info("Parsing filenames")
-    stored = pysat.utils.files.parse_fixed_width_filenames(filenames,
-                                                           format_str)
+    if format_str.find('*') < 0:
+        stored = pysat.utils.files.parse_fixed_width_filenames(filenames,
+                                                               format_str)
+    else:
+        stored = pysat.utils.files.parse_delimited_filenames(filenames,
+                                                             format_str, '.')
 
     # Process the parsed filenames and return a properly formatted Series
     logger.info("Processing filenames")
@@ -727,10 +1201,10 @@ def list_remote_files(tag, inst_id, inst_code=None, kindats=None, user=None,
                                                       two_digit_year_break)
 
 
-def list_files(tag, inst_id, data_path=None, format_str=None,
+def list_files(tag, inst_id, data_path, format_str=None,
                supported_tags=None, file_cadence=dt.timedelta(days=1),
                two_digit_year_break=None, delimiter=None, file_type=None):
-    """Return a Pandas Series of every file for chosen Instrument data.
+    """Create a Pandas Series of every file for chosen Instrument data.
 
     Parameters
     ----------
@@ -740,9 +1214,8 @@ def list_files(tag, inst_id, data_path=None, format_str=None,
     inst_id : str
         Specifies the instrument ID to load. Accepts strings corresponding to
         the appropriate Madrigal Instrument `inst_ids`.
-    data_path : str or NoneType
-        Path to data directory.  If None is specified, the value previously
-        set in Instrument.files.data_path is used.  (default=None)
+    data_path : str
+        Path to data directory.
     format_str : str or NoneType
         User specified file format.  If None is specified, the default
         formats associated with the supplied tags are used. (default=None)
@@ -876,13 +1349,19 @@ def _check_madrigal_params(inst_code, user, password):
 
     """
 
-    if inst_code is None:
-        raise ValueError("Must supply Madrigal instrument code")
+    inst_codes = known_madrigal_inst_codes(None)
+
+    if str(inst_code) not in inst_codes.keys():
+        raise ValueError(''.join(["Unknown Madrigal instrument code: ",
+                                  repr(inst_code), ". If this is a valid ",
+                                  "Madrigal instrument code, please update ",
+                                  "`pysatMadrigal.instruments.methods.general",
+                                  ".known_madrigal_inst_codes`."]))
 
     if not (isinstance(user, str) and isinstance(password, str)):
         raise ValueError(' '.join(("The madrigal database requries a username",
                                    "and password.  Please input these as",
-                                   "user='firstname+lastname' and",
+                                   "user='firstname lastname' and",
                                    "password='myname@email.address' in this",
                                    "function.")))